Made changes to use npm to manage dependancies and remove modules from git repository.
This included adding a basic package.json with dependancies listed
(this will need to be updated with correct details). Also added the
'private' flag to prevent accidentally publishing the code.

Also added the 'node_modules/' directory to the .gitignore to prevent
future commits of node modules added to the project.